+++ packages/cli/src/scaffold/ethereum.test.ts
@@ -82,7 +82,9 @@ const scaffoldWithIndexEvents = new Scaffold({
 describe.concurrent('Ethereum subgraph scaffolding', () => {
   test('Manifest', async () => {
     expect(await scaffold.generateManifest()).toEqual(`\
-specVersion: 0.0.5
+specVersion: 1.0.0
+indexerHints:
+  prune: auto
 schema:
   file: ./schema.graphql
 dataSources: