+++ packages/ts/chain/ethereum.ts
@@ -5,6 +5,8 @@ import { Address, BigInt } from '../common/numbers';
 /** Host Ethereum interface */
 export declare namespace ethereum {
   function call(call: SmartContractCall): Array<Value> | null;
+  function getBalance(address: Address): BigInt;
+  function hasCode(address: Address): Wrapped<bool>;
   function encode(token: Value): Bytes | null;
   function decode(types: String, data: Bytes): Value | null;
 }